Refactor shared logic into payment flow managing protocol
