+++ Sources/PrimerSDK/Classes/Core/PrimerHeadlessUniversalCheckout/Composable/Klarna/Components/PrimerHeadlessKlarnaComponent+SessionCreation.swift
@@ -120,44 +120,5 @@ extension PrimerHeadlessKlarnaComponent: LogReporter {
         }
         handleReceivedError(error: primerError)
     }
-
-    func handlePrimerWillCreatePaymentEvent(_ paymentMethodData: PrimerPaymentMethodData) -> Promise<Void> {
-        return Promise { seal in
-            if PrimerInternal.shared.intent == .vault {
-                seal.fulfill()
-            } else {
-                let checkoutPaymentMethodType = PrimerCheckoutPaymentMethodType(type: paymentMethodData.type)
-                let checkoutPaymentMethodData = PrimerCheckoutPaymentMethodData(type: checkoutPaymentMethodType)
-
-                var decisionHandlerHasBeenCalled = false
-
-                PrimerDelegateProxy.primerWillCreatePaymentWithData(
-                    checkoutPaymentMethodData,
-                    decisionHandler: { paymentCreationDecision in
-                        decisionHandlerHasBeenCalled = true
-                        switch paymentCreationDecision.type {
-                        case .abort(let errorMessage):
-                            let error = PrimerError.merchantError(message: errorMessage ?? "",
-                                                                  userInfo: .errorUserInfoDictionary(),
-                                                                  diagnosticsId: UUID().uuidString)
-                            seal.reject(error)
-                        case .continue:
-                            seal.fulfill()
-                        }
-                    })
-
-                DispatchQueue.main.asyncAfter(deadline: .now() + 5) { [weak self] in
-                    if !decisionHandlerHasBeenCalled {
-                        let message =
-                            """
-The 'decisionHandler' of 'primerHeadlessUniversalCheckoutWillCreatePaymentWithData' hasn't been called. \
-Make sure you call the decision handler otherwise the SDK will hang.
-"""
-                        self?.logger.warn(message: message)
-                    }
-                }
-            }
-        }
-    }
 }
 #endif
