+++ Sources/PrimerSDK/Classes/Core/PrimerHeadlessUniversalCheckout/Managers/Payment Method Managers/RawDataManager.swift
@@ -53,7 +53,7 @@ protocol PrimerRawDataTokenizationBuilderProtocol {
 
 extension PrimerHeadlessUniversalCheckout {
 
-    public class RawDataManager: NSObject, LogReporter {
+    public class RawDataManager: NSObject, PaymentFlowManaging, LogReporter {
 
         public var delegate: PrimerHeadlessUniversalCheckoutRawDataManagerDelegate?
         public private(set) var paymentMethodType: String
@@ -78,8 +78,8 @@ extension PrimerHeadlessUniversalCheckout {
 
             return self.rawDataTokenizationBuilder.requiredInputElementTypes
         }
-        private var resumePaymentId: String?
-        public private(set) var paymentCheckoutData: PrimerCheckoutData?
+        var resumePaymentId: String?
+        var paymentCheckoutData: PrimerCheckoutData?
         public private(set) var isDataValid: Bool = false
         var webViewController: SFSafariViewController?
         private var webViewCompletion: ((_ authorizationToken: String?, _ error: PrimerError?) -> Void)?
@@ -267,46 +267,7 @@ extension PrimerHeadlessUniversalCheckout {
             return (rawDataTokenizationBuilder as? PrimerRawCardDataTokenizationBuilder)?
                 .validateRawData(rawData, cardNetworksMetadata: cardNetworksMetadata)
         }
-
-        private func handlePrimerWillCreatePaymentEvent(_ paymentMethodData: PrimerPaymentMethodData) -> Promise<Void> {
-            return Promise { seal in
-                if PrimerInternal.shared.intent == .vault {
-                    seal.fulfill()
-                } else {
-                    let checkoutPaymentMethodType = PrimerCheckoutPaymentMethodType(type: paymentMethodData.type)
-                    let checkoutPaymentMethodData = PrimerCheckoutPaymentMethodData(type: checkoutPaymentMethodType)
-
-                    var decisionHandlerHasBeenCalled = false
-
-                    PrimerDelegateProxy.primerWillCreatePaymentWithData(
-                        checkoutPaymentMethodData,
-                        decisionHandler: { paymentCreationDecision in
-                            decisionHandlerHasBeenCalled = true
-                            switch paymentCreationDecision.type {
-                            case .abort(let errorMessage):
-                                let error = PrimerError.merchantError(message: errorMessage ?? "",
-                                                                      userInfo: .errorUserInfoDictionary(),
-                                                                      diagnosticsId: UUID().uuidString)
-                                seal.reject(error)
-                            case .continue:
-                                seal.fulfill()
-                            }
-                        })
-
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 5) { [weak self] in
-                        if !decisionHandlerHasBeenCalled {
-                            let message =
-                                """
-"The 'decisionHandler' of 'primerHeadlessUniversalCheckoutWillCreatePaymentWithData' hasn't been called. \
-Make sure you call the decision handler otherwise the SDK will hang."
-"""
-                            self?.logger.warn(message: message)
-                        }
-                    }
-                }
-            }
-        }
-
+        
         private func makeRequestBody() -> Promise<Request.Body.Tokenization> {
 
             return Promise { seal in
@@ -363,142 +324,6 @@ Make sure you call the decision handler otherwise the SDK will hang."
                 }
                 .catch { err in
                     seal.reject(err)
-                }
-            }
-        }
-
-        // This function will do one of the two following:
-        //     - Wait a response from the merchant, via the delegate function. The response can be:
-        //         - A new client token
-        //         - Success
-        //         - Error
-        //     - Perform the payment internally, and get a response from our BE. The response will
-        //       be a Payment response. The can contain:
-        //         - A required action with a new client token
-        //         - Be successful
-        //         - Has failed
-        //
-        // Therefore, return:
-        //     - A decoded client token
-        //     - nil for success
-        //     - Reject with an error
-        private func startPaymentFlowAndFetchDecodedClientToken(withPaymentMethodTokenData paymentMethodTokenData: PrimerPaymentMethodTokenData) -> Promise<DecodedJWTToken?> {
-            return Promise { seal in
-                if PrimerSettings.current.paymentHandling == .manual {
-                    PrimerDelegateProxy.primerDidTokenizePaymentMethod(paymentMethodTokenData) { resumeDecision in
-                        if let resumeType = resumeDecision.type as? PrimerResumeDecision.DecisionType {
-                            switch resumeType {
-                            case .succeed:
-                                seal.fulfill(nil)
-
-                            case .continueWithNewClientToken(let newClientToken):
-                                let apiConfigurationModule = PrimerAPIConfigurationModule()
-
-                                firstly {
-                                    apiConfigurationModule.storeRequiredActionClientToken(newClientToken)
-                                }
-                                .done {
-                                    guard let decodedJWTToken = PrimerAPIConfigurationModule.decodedJWTToken else {
-                                        let err = PrimerError.invalidClientToken(userInfo: .errorUserInfoDictionary(),
-                                                                                 diagnosticsId: UUID().uuidString)
-                                        ErrorHandler.handle(error: err)
-                                        throw err
-                                    }
-
-                                    seal.fulfill(decodedJWTToken)
-                                }
-                                .catch { err in
-                                    seal.reject(err)
-                                }
-
-                            case .fail(let message):
-                                var merchantErr: Error!
-                                if let message = message {
-                                    let err = PrimerError.merchantError(message: message,
-                                                                        userInfo: .errorUserInfoDictionary(),
-                                                                        diagnosticsId: UUID().uuidString)
-                                    merchantErr = err
-                                } else {
-                                    merchantErr = NSError.emptyDescriptionError
-                                }
-                                seal.reject(merchantErr)
-                            }
-
-                        } else if let resumeDecisionType = resumeDecision.type as? PrimerHeadlessUniversalCheckoutResumeDecision.DecisionType {
-                            switch resumeDecisionType {
-                            case .continueWithNewClientToken(let newClientToken):
-                                let apiConfigurationModule: PrimerAPIConfigurationModuleProtocol = PrimerAPIConfigurationModule()
-
-                                firstly {
-                                    apiConfigurationModule.storeRequiredActionClientToken(newClientToken)
-                                }
-                                .done {
-                                    guard let decodedJWTToken = PrimerAPIConfigurationModule.decodedJWTToken else {
-                                        let err = PrimerError.invalidClientToken(userInfo: .errorUserInfoDictionary(),
-                                                                                 diagnosticsId: UUID().uuidString)
-                                        ErrorHandler.handle(error: err)
-                                        throw err
-                                    }
-
-                                    seal.fulfill(decodedJWTToken)
-                                }
-                                .catch { err in
-                                    seal.reject(err)
-                                }
-
-                            case .complete:
-                                seal.fulfill(nil)
-                            }
-
-                        } else {
-                            precondition(false)
-                        }
-                    }
-
-                } else {
-                    guard let token = paymentMethodTokenData.token else {
-                        let err = PrimerError.invalidClientToken(
-                            userInfo: .errorUserInfoDictionary(),
-                            diagnosticsId: UUID().uuidString)
-                        ErrorHandler.handle(error: err)
-                        seal.reject(err)
-                        return
-                    }
-
-                    firstly {
-                        self.handleCreatePaymentEvent(token)
-                    }
-                    .done { paymentResponse -> Void in
-                        self.paymentCheckoutData = PrimerCheckoutData(payment: PrimerCheckoutDataPayment(from: paymentResponse))
-                        self.resumePaymentId = paymentResponse.id
-
-                        if let requiredAction = paymentResponse.requiredAction {
-                            let apiConfigurationModule = PrimerAPIConfigurationModule()
-
-                            firstly {
-                                apiConfigurationModule.storeRequiredActionClientToken(requiredAction.clientToken)
-                            }
-                            .done {
-                                guard let decodedJWTToken = PrimerAPIConfigurationModule.decodedJWTToken else {
-                                    let err = PrimerError.invalidClientToken(userInfo: .errorUserInfoDictionary(),
-                                                                             diagnosticsId: UUID().uuidString)
-                                    ErrorHandler.handle(error: err)
-                                    throw err
-                                }
-
-                                seal.fulfill(decodedJWTToken)
-                            }
-                            .catch { err in
-                                seal.reject(err)
-                            }
-
-                        } else {
-                            seal.fulfill(nil)
-                        }
-                    }
-                    .catch { err in
-                        seal.reject(err)
-                    }
                 }
             }
         }
@@ -720,82 +545,6 @@ Make sure you call the decision handler otherwise the SDK will hang."
                     seal.reject(err)
                 }
             }
-        }
-
-        private func handleResumeStepsBasedOnSDKSettings(resumeToken: String) -> Promise<PrimerCheckoutData?> {
-            return Promise { seal in
-                if PrimerSettings.current.paymentHandling == .manual {
-                    PrimerDelegateProxy.primerDidResumeWith(resumeToken) { resumeDecision in
-                        if let resumeDecisionType = resumeDecision.type as? PrimerResumeDecision.DecisionType {
-                            switch resumeDecisionType {
-                            case .fail(let message):
-                                var merchantErr: Error!
-                                if let message = message {
-                                    let err = PrimerError.merchantError(message: message,
-                                                                        userInfo: .errorUserInfoDictionary(),
-                                                                        diagnosticsId: UUID().uuidString)
-                                    merchantErr = err
-                                } else {
-                                    merchantErr = NSError.emptyDescriptionError
-                                }
-                                seal.reject(merchantErr)
-
-                            case .succeed:
-                                seal.fulfill(nil)
-
-                            case .continueWithNewClientToken:
-                                seal.fulfill(nil)
-                            }
-
-                        } else if let resumeDecisionType = resumeDecision.type as? PrimerHeadlessUniversalCheckoutResumeDecision.DecisionType {
-                            switch resumeDecisionType {
-                            case .continueWithNewClientToken:
-                                seal.fulfill(self.paymentCheckoutData)
-
-                            case .complete:
-                                seal.fulfill(self.paymentCheckoutData)
-                            }
-
-                        } else {
-                            precondition(false)
-                        }
-                    }
-
-                } else {
-                    guard let resumePaymentId = self.resumePaymentId else {
-                        let resumePaymentIdError = PrimerError.invalidValue(key: "resumePaymentId",
-                                                                            value: "Resume Payment ID not valid",
-                                                                            userInfo: .errorUserInfoDictionary(),
-                                                                            diagnosticsId: UUID().uuidString)
-                        ErrorHandler.handle(error: resumePaymentIdError)
-                        seal.reject(resumePaymentIdError)
-                        return
-                    }
-
-                    firstly {
-                        self.handleResumePaymentEvent(resumePaymentId, resumeToken: resumeToken)
-                    }
-                    .done { paymentResponse -> Void in
-                        let paymentData = PrimerCheckoutDataPayment(from: paymentResponse)
-                        self.paymentCheckoutData = PrimerCheckoutData(payment: paymentData)
-                        seal.fulfill(self.paymentCheckoutData)
-                    }
-                    .catch { err in
-                        seal.reject(err)
-                    }
-                }
-            }
-        }
-
-        private func handleCreatePaymentEvent(_ paymentMethodData: String) -> Promise<Response.Body.Payment> {
-            let paymentRequest = Request.Body.Payment.Create(token: paymentMethodData)
-            return createResumePaymentService.createPayment(paymentRequest: paymentRequest)
-        }
-
-        private func handleResumePaymentEvent(_ resumePaymentId: String, resumeToken: String) -> Promise<Response.Body.Payment> {
-            let resumeRequest = Request.Body.Payment.Resume(token: resumeToken)
-            return createResumePaymentService.resumePaymentWithPaymentId(resumePaymentId,
-                                                                         paymentResumeRequest: resumeRequest)
         }
 
         private func presentWebRedirectViewControllerWithRedirectUrl(_ redirectUrl: URL) -> Promise<Void> {
