+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/PaymentMethodTokenizationViewModel+Logic.swift
@@ -12,7 +12,7 @@
 import Foundation
 import UIKit
 
-extension PaymentMethodTokenizationViewModel {
+extension PaymentMethodTokenizationViewModel: PaymentFlowManaging {
 
     @objc
     func start() {
@@ -239,206 +239,6 @@ extension PaymentMethodTokenizationViewModel {
         }
     }
 
-    // This function will do one of the two following:
-    //     - Wait a response from the merchant, via the delegate function. The response can be:
-    //         - A new client token
-    //         - Success
-    //         - Error
-    //     - Perform the payment internally, and get a response from our BE. The response will
-    //       be a Payment response. The can contain:
-    //         - A required action with a new client token
-    //         - Be successful
-    //         - Has failed
-    //
-    // Therefore, return:
-    //     - A decoded client token
-    //     - nil for success
-    //     - Reject with an error
-
-    func startPaymentFlowAndFetchDecodedClientToken(withPaymentMethodTokenData paymentMethodTokenData: PrimerPaymentMethodTokenData) -> Promise<DecodedJWTToken?> {
-        return Promise { seal in
-            if PrimerSettings.current.paymentHandling == .manual {
-                PrimerDelegateProxy.primerDidTokenizePaymentMethod(paymentMethodTokenData) { resumeDecision in
-                    if let resumeDecisionType = resumeDecision.type as? PrimerResumeDecision.DecisionType {
-                        switch resumeDecisionType {
-                        case .succeed:
-                            seal.fulfill(nil)
-
-                        case .continueWithNewClientToken(let newClientToken):
-                            let apiConfigurationModule = PrimerAPIConfigurationModule()
-
-                            firstly {
-                                apiConfigurationModule.storeRequiredActionClientToken(newClientToken)
-                            }
-                            .done {
-                                guard let decodedJWTToken = PrimerAPIConfigurationModule.decodedJWTToken else {
-                                    let err = PrimerError.invalidClientToken(userInfo: .errorUserInfoDictionary(),
-                                                                             diagnosticsId: UUID().uuidString)
-                                    ErrorHandler.handle(error: err)
-                                    throw err
-                                }
-
-                                seal.fulfill(decodedJWTToken)
-                            }
-                            .catch { err in
-                                seal.reject(err)
-                            }
-
-                        case .fail(let message):
-                            var merchantErr: Error!
-                            if let message = message {
-                                let err = PrimerError.merchantError(message: message,
-                                                                    userInfo: .errorUserInfoDictionary(),
-                                                                    diagnosticsId: UUID().uuidString)
-                                merchantErr = err
-                            } else {
-                                merchantErr = NSError.emptyDescriptionError
-                            }
-                            seal.reject(merchantErr)
-                        }
-
-                    } else if let resumeDecisionType = resumeDecision.type as? PrimerHeadlessUniversalCheckoutResumeDecision.DecisionType {
-                        switch resumeDecisionType {
-                        case .continueWithNewClientToken(let newClientToken):
-                            let apiConfigurationModule: PrimerAPIConfigurationModuleProtocol = PrimerAPIConfigurationModule()
-
-                            firstly {
-                                apiConfigurationModule.storeRequiredActionClientToken(newClientToken)
-                            }
-                            .done {
-                                guard let decodedJWTToken = PrimerAPIConfigurationModule.decodedJWTToken else {
-                                    let err = PrimerError.invalidClientToken(userInfo: .errorUserInfoDictionary(),
-                                                                             diagnosticsId: UUID().uuidString)
-                                    ErrorHandler.handle(error: err)
-                                    throw err
-                                }
-
-                                seal.fulfill(decodedJWTToken)
-                            }
-                            .catch { err in
-                                seal.reject(err)
-                            }
-
-                        case .complete:
-                            seal.fulfill(nil)
-                        }
-
-                    } else {
-                        precondition(false)
-                    }
-                }
-
-            } else {
-                guard let token = paymentMethodTokenData.token else {
-                    let err = PrimerError.invalidClientToken(
-                        userInfo: .errorUserInfoDictionary(),
-                        diagnosticsId: UUID().uuidString)
-                    ErrorHandler.handle(error: err)
-                    seal.reject(err)
-                    return
-                }
-
-                firstly {
-                    self.handleCreatePaymentEvent(token)
-                }
-                .done { paymentResponse -> Void in
-                    self.paymentCheckoutData = PrimerCheckoutData(payment: PrimerCheckoutDataPayment(from: paymentResponse))
-                    self.resumePaymentId = paymentResponse.id
-
-                    if let requiredAction = paymentResponse.requiredAction {
-                        let apiConfigurationModule = PrimerAPIConfigurationModule()
-
-                        firstly {
-                            apiConfigurationModule.storeRequiredActionClientToken(requiredAction.clientToken)
-                        }
-                        .done {
-                            guard let decodedJWTToken = PrimerAPIConfigurationModule.decodedJWTToken else {
-                                let err = PrimerError.invalidClientToken(userInfo: .errorUserInfoDictionary(),
-                                                                         diagnosticsId: UUID().uuidString)
-                                ErrorHandler.handle(error: err)
-                                throw err
-                            }
-
-                            seal.fulfill(decodedJWTToken)
-                        }
-                        .catch { err in
-                            seal.reject(err)
-                        }
-
-                    } else {
-                        seal.fulfill(nil)
-                    }
-                }
-                .catch { err in
-                    seal.reject(err)
-                }
-            }
-        }
-    }
-
-    func handleResumeStepsBasedOnSDKSettings(resumeToken: String) -> Promise<PrimerCheckoutData?> {
-        return Promise { seal in
-            if PrimerSettings.current.paymentHandling == .manual {
-                PrimerDelegateProxy.primerDidResumeWith(resumeToken) { resumeDecision in
-                    if let resumeDecisionType = resumeDecision.type as? PrimerResumeDecision.DecisionType {
-                        switch resumeDecisionType {
-                        case .fail(let message):
-                            var merchantErr: Error!
-                            if let message = message {
-                                let err = PrimerError.merchantError(message: message,
-                                                                    userInfo: .errorUserInfoDictionary(),
-                                                                    diagnosticsId: UUID().uuidString)
-                                merchantErr = err
-                            } else {
-                                merchantErr = NSError.emptyDescriptionError
-                            }
-                            seal.reject(merchantErr)
-
-                        case .succeed:
-                            seal.fulfill(nil)
-
-                        case .continueWithNewClientToken:
-                            seal.fulfill(nil)
-                        }
-
-                    } else if let resumeDecisionType = resumeDecision.type as? PrimerHeadlessUniversalCheckoutResumeDecision.DecisionType {
-                        switch resumeDecisionType {
-                        case .continueWithNewClientToken:
-                            seal.fulfill(nil)
-                        case .complete:
-                            seal.fulfill(nil)
-                        }
-
-                    } else {
-                        precondition(false)
-                    }
-                }
-
-            } else {
-                guard let resumePaymentId = self.resumePaymentId else {
-                    let resumePaymentIdError = PrimerError.invalidValue(key: "resumePaymentId",
-                                                                        value: "Resume Payment ID not valid",
-                                                                        userInfo: .errorUserInfoDictionary(),
-                                                                        diagnosticsId: UUID().uuidString)
-                    ErrorHandler.handle(error: resumePaymentIdError)
-                    seal.reject(resumePaymentIdError)
-                    return
-                }
-
-                firstly {
-                    self.handleResumePaymentEvent(resumePaymentId, resumeToken: resumeToken)
-                }
-                .done { paymentResponse -> Void in
-                    self.paymentCheckoutData = PrimerCheckoutData(payment: PrimerCheckoutDataPayment(from: paymentResponse))
-                    seal.fulfill(self.paymentCheckoutData)
-                }
-                .catch { err in
-                    seal.reject(err)
-                }
-            }
-        }
-    }
-
     func handleFailureFlow(errorMessage: String?) {
         let categories = self.config.paymentMethodManagerCategories
         PrimerUIManager.dismissOrShowResultScreen(
@@ -446,59 +246,6 @@ extension PaymentMethodTokenizationViewModel {
             paymentMethodManagerCategories: categories ?? [],
             withMessage: errorMessage
         )
-    }
-
-    internal func handlePrimerWillCreatePaymentEvent(_ paymentMethodData: PrimerPaymentMethodData) -> Promise<Void> {
-        return Promise { seal in
-            if PrimerInternal.shared.intent == .vault {
-                seal.fulfill()
-            } else {
-                let checkoutPaymentMethodType = PrimerCheckoutPaymentMethodType(type: paymentMethodData.type)
-                let checkoutPaymentMethodData = PrimerCheckoutPaymentMethodData(type: checkoutPaymentMethodType)
-
-                var decisionHandlerHasBeenCalled = false
-
-                PrimerDelegateProxy.primerWillCreatePaymentWithData(
-                    checkoutPaymentMethodData,
-                    decisionHandler: { paymentCreationDecision in
-                        decisionHandlerHasBeenCalled = true
-                        switch paymentCreationDecision.type {
-                        case .abort(let errorMessage):
-                            let error = PrimerError.merchantError(message: errorMessage ?? "",
-                                                                  userInfo: .errorUserInfoDictionary(),
-                                                                  diagnosticsId: UUID().uuidString)
-                            seal.reject(error)
-                        case .continue:
-                            seal.fulfill()
-                        }
-                    })
-
-                DispatchQueue.main.asyncAfter(deadline: .now() + 5) { [weak self] in
-                    if !decisionHandlerHasBeenCalled {
-                        let message =
-                            """
-The 'decisionHandler' of 'primerHeadlessUniversalCheckoutWillCreatePaymentWithData' hasn't been called. \
-Make sure you call the decision handler otherwise the SDK will hang.
-"""
-                        self?.logger.warn(message: message)
-                    }
-                }
-            }
-        }
-    }
-
-    // Create payment with Payment method token
-
-    private func handleCreatePaymentEvent(_ paymentMethodData: String) -> Promise<Response.Body.Payment> {
-        let body = Request.Body.Payment.Create(token: paymentMethodData)
-        return createResumePaymentService.createPayment(paymentRequest: body)
-    }
-
-    // Resume payment with Resume payment ID
-
-    private func handleResumePaymentEvent(_ resumePaymentId: String, resumeToken: String) -> Promise<Response.Body.Payment> {
-        let body = Request.Body.Payment.Resume(token: resumeToken)
-        return createResumePaymentService.resumePaymentWithPaymentId(resumePaymentId, paymentResumeRequest: body)
     }
 
     func validateReturningPromise() -> Promise<Void> {
